Update patch set 6
Patch Set 6: Code-Review+1

Sorry for not noticing this earlier, but this is for the OpenGL ES front-end? We're in the process of deprecating that, in favor of using ANGLE on top of the Vulkan implementation (SwANGLE).

We'll keep the legacy GLES implementation around as long as necessary, but are getting close to being able to remove it for Chrome and Android use cases.

Can you focus efforts on getting SwANGLE to work for Fuchsia, or is this needed in the short term? Fine with me to land it if Antonio or Ben can review the CMake changes.

Patch-set: 6
Reviewer: Gerrit User 9669 <9669@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+1