Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Regretfully, I think I will have to abandon this CL. The reason for this is that the CMake build will break as soon as https://swiftshader-review.googlesource.com/c/SwiftShader/+/37908 lands, because it uses Fuchsia FIDL client libraries (e.g. with code like #include <fuchsia/sysmem/cpp/fidl.h>), required to access system services at runtime.
> 
> And unfortunately, these libraries are _not_ provided by the Fuchsia Core SDK. Instead the SDK provides .fidl files that needs to be parsed by host tools like "fidlc" and "fidlgen" to generate corresponding headers, C++ sources files. The latter which also need to be manually compiled through CMake and statically linked into the SwiftShader binary.
> 
> There are no CMake modules / rules to do that at the moment, and implementing these properly is a rather difficult task (and these would have to be provided out of SwiftShader somewhere).
> 
> As such, I think I will only support building SwiftShader for Fuchsia with the Chromium build, or the new Fuchsia GN SDK (https://fuchsia.dev/fuchsia-src/development/sdk/gn) which is also GN-based and supports properly using FIDL libraries out of the box.
> 
> If I ever take the time to implement scripts to implement the proper CMake modules for all of the FIDL libraries in the Core SDK, I'll reopen this CL with appropriate updates.

And I must recant what I wrote earlier :-), because it turns out that trying to use the GN-based Fuchsia SDK is impossible without drastically changing the current BUILD.gn.

Instead, I found a way to generate a minimalistic CMake-based Fuchsia SDK instead [1], allowing me to use cmake to build SwiftShader, with support for Fuchsia SDK package and FIDL libraries (think of them as the equivalent of system libraries).

So here's a new patch for this CL. Only building the Vulkan ICD is supported, since there is no need for EGL/GLES libraries on Fuchsia, so this won't interface with your SWAngle work.

Sorry for the noise. I now have a SwiftShader ICD that runs inside a Fuchsia process. Still many things to do to make it usable, but that's a big step for us :)

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/379673

Patch-set: 6