vk::Device: Use SyncCache for SamplingRoutineCache
Replace the logic of `getOrCreate()` with `SyncCache`.

This will ensure that the routine will only be created once even if multiple concurrent calls are made with the same key. Threads waiting for the entry to be created may yield, allowing the thread to do other work instead of blocking until the entry is created.

Bug: b/154138414
Change-Id: Id0fd39de2426326cdef7d0425c87c645f62cb346