vk::PipelineCache: Use SyncCache.
Replace the logic of `getOrCreateComputeProgram()` and `getOrCreateShader()` with `SyncCache`.

This will ensure that the program and shader will only be created once even if multiple concurrent calls are made with the same key. Threads waiting for the entry to be created may yield, allowing the thread to do other work instead of blocking until the entry is created.

Bug: TODO
Change-Id: I99ba09de4fe7981da23aab3b8a9de2c2b4582e1f