VkPipeline: Background compile shaders.
Use a `SyncCache` to hold the compiled shaders / program.
This serves two purposes:
(1) We can call `SyncCache::getOrCreate` on a separate marl task to background compile. If this hasn't finished compiling by the time we try to use it, it'll block.
(2) The SyncCache key (CompileOptions) gives us a foundation for dynamically rebuilding shaders with different flags. Examples:
* Different optimization levels
* Different reactor backends.
* Enabling / disabing debugger features.

Bug: TODO
Change-Id: I5f20219e7742d443d8d127dc7ea7c195107773c0