Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I'm not sure this needs to land. As previously discussed I'd like to avoid having a tools/ directory that's a catch-all for various scripts that may at one point be useful to someone but have little lasting value.

Hmm, although I understand not wanting a catch-all pool of scripts, we already have something like this when it comes to stuff like Kokoro scripts, and scripts for emitting coverage information.

In this specific case, it's useful to see the dependency graph of our targets to get a better understanding of the overall structure. I was even thinking of adding a script to Kokoro to run this and update a png on our g3doc site. Perhaps we just need a better folder structure to store this in?

Patch-set: 4
Reviewer: Gerrit User 9669 <9669@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>