CMake: split out test and benchmark targets
Moved out code from the root CMakeLists into new CMakeLists for each
test and benchmark directory. For Reactor, both the unit test and
benchmark source files are in src/Reactor, so for now, I created folders
for these under tests with a CMakeLists that includes the source files
from src/Reactor. This will be cleaned up eventually when we move
Reactor out of src (b/154013190).

Bug: b/145758253
Change-Id: Ic6bbc394e49d3bdc3105509a19458722fcfe68ee