Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1: Code-Review+2
> 
> Given that we're now always importing the subzero CMakeLists.txt even if we use it or not, I question whether we should even warn if subzero isn't supported by that platform. Seems unnecessarily noisy for perfectly valid targets.
> 
> Alternatively, I suppose we could promote the FATAL_ERROR up to the root CMakeLists.txt where it has the responsibility of deciding which backend is being used.
> 
> Anyway, I don't want to block a change that fixes a build, so +2 from me. Antonio and I can discuss followup changes in our own time.

Yes, let's go with making this a warning for now. Ultimately, if we do the tiered JIT thing, we'll want to add both targets like this, and then depending on which targets were successfully added, we can support tiered or not. In the future, rather than a warning, we'll probably make it silent, and expect the parent project to check if the target was created (or maybe set a parent-scope variable).

Patch-set: 1
Reviewer: Gerrit User 9669 <9669@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+2