Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> Be aware that we will want to reduce the enabled config.h features down to something similar to llvm-7's. I had a quick stab at this, but it was quite fiddly and more urgent things came up. Your processes currently matches mine for Linux (and macOS, but I haven't landed that change yet due to b/145545790)
> 
> Once unnecessary features are turned off, we can use build/strip_unneeded.go (https://swiftshader.googlesource.com/SwiftShader/+/refs/heads/master/build/strip_unneeded.go) again to churn through the llvm source files and remove the source files that are unnecessarily built.

Yes, I'm aware that this is of limited use, because I've tried to rebuild SwiftShader/Fuchsia/arm64 with LLVM 10.0, and the llvm-config.h provided here definitely only work for x86/x86_64 binaries.

It's good to know that you have a plan to attach this. I assume you will end up writing a script similar to third_party/llvm-7.0/scripts/update-llvm.py to recompute the right content of this header file. Let me know if you need a reviewer for something like this (oh, and SwiftShader/Fuchsia/arm64 is not a priority right now).

Patch-set: 2