Update patch set 17
Patch Set 17:

> Patch Set 17:
> 
> I think it's great to have thread safety analysis enabled, but I am not a fan of having all the thread primitive types come from the marl namespace. The fact that marl offers TSA wrappers, along with the CMake check, is convenient, but feels out of place. When I see "marl" in our codebase, I expect to see code dealing with task scheduling. I would much prefer if the TSA wrappers and CMakeLists were a separate library. I understand that it's a pain to open source stuff at Google, but there must be some kind of middle ground?

Why would coming out of a separate library be any better?

Marl owns all the threads used by SwiftShader, and was already providing a replacement for std::condition_variable. The use of any std synchronization primitive is going to block the marl worker threads, causing pipeline bubbles. If there are places using std:: synchronization primitives, they should be replaced.

So I think it's clear to say marl is a task scheduler and thread library.

Marl was born in SwiftShader, and was only separated out because there was interest from Android.
SwiftShader is still Marl's highest priority client. It's not going to go breaking things in ways that's not serving SwiftShader's needs.

I don't understand this desire to break it up into smaller libraries.

Patch-set: 17