Update patch set 17
Patch Set 17:

> Patch Set 17:
> 
> > Patch Set 17:
> > 
> > > Patch Set 17:
> > > 
> > > I think it's great to have thread safety analysis enabled, but I am not a fan of having all the thread primitive types come from the marl namespace. The fact that marl offers TSA wrappers, along with the CMake check, is convenient, but feels out of place. When I see "marl" in our codebase, I expect to see code dealing with task scheduling. I would much prefer if the TSA wrappers and CMakeLists were a separate library. I understand that it's a pain to open source stuff at Google, but there must be some kind of middle ground?
> > 
> > Why would coming out of a separate library be any better?
> > 
> > Marl owns all the threads used by SwiftShader, and was already providing a replacement for std::condition_variable. The use of any std synchronization primitive is going to block the marl worker threads, causing pipeline bubbles. If there are places using std:: synchronization primitives, they should be replaced.
> > 
> > So I think it's clear to say marl is a task scheduler and thread library.
> 
> Fair enough. Marl started off as a task scheduling library, but it's becoming more of a general thread library, as you point out. Might be good to update the readme for Marl to say as much.
> 
> > Marl was born in SwiftShader, and was only separated out because there was interest from Android.
> > SwiftShader is still Marl's highest priority client. It's not going to go breaking things in ways that's not serving SwiftShader's needs.
> > I don't understand this desire to break it up into smaller libraries.
> 
> I'm not worried about Marl breaking SwiftShader. It's more about separation of concerns, and the usual worry that useful utilities end up being mixed in. For example, "defer" is useful, and has nothing to do with threads, but to use it, we include "marl/defer.h".
> 
> In this specific case, for TSA, I was thinking how generally useful it would be to have simple wrappers around the existing std thread primitives that do nothing more than wrap them with TSA macros. It would be great to be able to replace std:: with tsa:: for these types, for example. However, I do realize that this is not a trivial thing to do, as there are many types to wrap, and many functions to forward; and perhaps more importantly, as you've pointed out before, TSA can't be applied to certain functions (e.g. can't have "lock/unlock" on mutex). So I suppose we can let this go. 

To be fair, the marl/tsa.h and marl/mutex.h headers are mostly the recommended wrappers taken from https://clang.llvm.org/docs/ThreadSafetyAnalysis.html#mutex-h, but tweaked a bit to fit with marl's types. If we feel really strongly, we could copy them out to SwiftShader, but I'm actively trying to migrate all the threading code over to marl, and this feels like a step in the wrong direction.

> Am starting to wonder just how useful TSA is now, given the limitations and difficulties with wrapping types, and the fact that we can use thread sanitizer.

It caught brokenness, so I'd say it is useful.

For what it's worth, nearly all of these annotations are removed in the changes I'm working on. I've replaced nearly every remaining use of std::mutex / marl::mutex with SyncCache (with the exception of the Vulkan debugger).

After thinking about this some more, pipeline bubbles aren't the worse thing that could happen with using the 'std' synchronization primitives. If you use these on a marl worker threads, you're potentially opening the door to deadlocks:

(1) Thread A is waiting on a std sync primitive, blocking the marl worker thread.
(2) Worker for thread A has a fiber queued that it cannot process due to (1) that will unblock task B
(3) Task B needs to be called to unblock the std sync primitive in (1), but it cannot.

Sort of thing that might cause a 1 in 100000 deadlock.

Patch-set: 17