Don't fail build if Vulkan headers are different
If a project that uses SwiftShader already imports the Vulkan::Headers
target, and these are different from the version in SwiftShader, we
fail to build because we have a static_assert to match the exact
version. Replace compile-time assert with a runtime assert to make
sure our code still works with different versions of the headers.

This fixes Amber building against SwiftShader.

Bug: b/154215163
Change-Id: I39e32370b51c702ab6d0a9790d06c31537f27472