Vulkan: Fix Fuchsia external semaphore fix tryWait().
Fix an issue where tryWait() would abort when called on an
unsignaled semaphore. Due to the fact that zx_object_wait_one()
will return ZX_ERR_TIMED_OUT in this case (because the expiration
time was set with zx_clock_get_monotonic() instead of
ZX_TIME_INFINITE).

This fixes an issue that appears when vkWaitForIdle() is called
by certain Fuchsia Vulkan graphics tests.

+ Replace ABORT() calls with DABORT() and proper returns to
  avoid crashing if DCHECK_ALWAYS_ON is not defined at
  build time.

Bug: None
Change-Id: Idaf14ed580c0c3ab326705842bcfe5e23eaad2b8