Vulkan: Add vk_icdInitializeConnectToServiceCallback()
This entry point is a private implementation detail between
the Fuchsia Vulkan loader and the Vulkan ICDs on this platform.

It is called by the loader early on to pass the address of a global
function pointer, that can later be used by extensions to connect
to Fuchsia FIDL services.

This is equivalent to the Fuchsia fdio_service_connect() function,
except that this scheme prevents adding libfdio as a dependency
into the ICD shared library.

Bug: None
Change-Id: Idd2a0a9b78495fff4bdb17d9f6afc446b067d7e0