CMake: fix astc code not being enabled correctly
In my refactoring of the astc-encoder (https://swiftshader-review.googlesource.com/c/SwiftShader/+/43269),
I had incorrectly set the dependency on astc-encoder in the Vulkan
module, rather than Device. Vulkan depends on Device, and we make use of
astc-encoder public macros in both modules. Thanks to bclayton@ for
finding and reporting this problem.

Fixes: b/154593186
Fixes: b/154593187
Fixes: b/154589883
Fixes: b/154591494
Fixes: b/154591735
Fixes: b/154593191
Fixes: b/154591498
Fixes: b/154593113
Change-Id: I8a71a6471ef85783110afb4f034a611116892b05