Don't unlock Android buffers immediately after locking them.
In some Android gralloc implementations (e.g. minigbm), the lock operation
corresponds to mmap, and unlock corresponds to munmap. This means that
this code was previously returning an unusable buffer in the case where
lockInternal is called passing LOCK_UNLOCKED (e.g. the mipmap buffer
initializer in Sampler.cpp), resulting in segfaults later on.

To prevent this from happening, don't unlock the native buffer and just leak
the reference.

Change-Id: I553801f32978c1d0af4597baad374381585e78ad