Fix specializing graphics routines for the pipeline layout
Shader modules can be used by different pipelines, even with incompatible
layouts. The generated code differs based on the layout, so we must use
it as part of the cache key.

This change conservatively assumes each pipeline layout to be unique, so
a serial ID is used to differentiate between them.

Bug: b/154660947
Change-Id: Id0dc5be9faaf137636ee2e27e6e84a7fbbf0010b