Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Just out of curiosity, how did this build fine before but produce black ASTC results? Is there a way to make it fail hard? Could affect refactorings of the other build files too.

Because all the ASTC logic is wrapped in `#ifdef SWIFTSHADER_ENABLE_ASTC`, and produces black images if it is not defined.
I'm guessing this is to avoid missing dependencies on some downstream builds?

I'm also curious to know why the offending change did not flag brokenness with regres. Will investigate.

Patch-set: 2