Eliminate the ABORT() macro
Terminating the process should be done with extreme caution. Usually
DABORT() is intended to be used instead. The ABORT() macro was making
it too easy to write code that causes the application to terminate in
relatively benign circumstances.

If scenarios are encountered which demand process termination (as
prescribed by the graphics API spec or as a platform requirement), we
can still use ::abort(), or bring back this macro with a scarier name.

Bug: b/154650520
Change-Id: I53780f72b22faadd62825d000661b605b77b597c