Fix ordering of descriptor set bindings
The Vulkan spec states that for vkCmdBindDescriptorSets():
"If any of the sets being bound include dynamic uniform or storage
buffers, then pDynamicOffsets includes one element for each array
element in each dynamic descriptor type binding in each set. Values are
taken from pDynamicOffsets in an order such that all entries for set N
come before set N+1; within a set, entries are ordered by the binding
numbers in the descriptor set layouts; and within a binding array,
elements are in order."

Ensuring that the offset values are associated with the right binding is
easiest to achieve when the entries are sorted by binding number.

In addition, the spec states that for vkUpdateDescriptorSets():
"If the dstBinding has fewer than descriptorCount array elements
remaining starting from dstArrayElement, then the remainder will be used
to update the subsequent binding - dstBinding+1 starting at array
element zero."

Which also demands that the bindings are sorted by binding number.

Bug: b/154241029
Change-Id: I1833e9b6ee2171bbdd8214d748b7a4ce211cac93