Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm all for this change, but it does mean that we should take a look at places where DABORT is used to make sure that we are actually able to continue running afterwards. Often times, cases where one aborts means it's not safe to keep executing afterwards, and will likely crash anyway, or worse, not crash and do funny things.

This has always been my argument for keeping this macro. If we know that things are so broken that things are going to crash and burn, let's put up a message and abort. Better to deterministically end the process with a clear message than knowingly let it go into a state that will likely crash, or worse still, expose security vulnerabilities.

I don't buy the argument that we must never terminate the process. It's not like we ever call ABORT() in code paths that we expect to be reached.

I do not support removing it.

Patch-set: 1
CC: Gerrit User 7183 <7183@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>