Update patch set 7
Patch Set 7:

Implementations have a lot of freedom in deciding when a helper invocation is active. My understanding is that this is because some hardware is able to clock gate or power gate SIMD lanes, or execute other active operations on them. For them it's beneficial to consider them logically inactive as much as reasonably possible. Other implementations just predicate the result at the bottom of the pipe, so they might as well consider helper invocations active all the time since the result is available anyway, could be "helpful", and there's no analysis to perform.

SwiftShader falls into the second camp currently, since SSE and NEON compute all the results for every SIMD lane and we just throw the logically inactive ones away with an explicit masking operation. This might change for AVX-512, which takes a mask registers as part of each instructions.

I'd rather not implement behavior that is counter to the 'physical' design. It's just one logical AND operation (and all the loads and stores to manage that extra mask), but it seems pretty clear to me that the result of OpGroupNonUniformBroadcastFirst is undefined for helper invocations. We're allowed to do whatever is cheapest.

Note that for OpGroupNonUniformQuad* it is well-defined that helper invocations must be active: "If a helper invocation would be active if it were not a helper invocation, it must be active for derivative and quad group operations." I think this also helps justify that one shouldn't be surprised if for OpGroupNonUniformBroadcastFirst they're active too.

Patch-set: 7
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>