SpirvShader: Document activeLaneMask() and storesAndAtomicsMask()
These had no documentation and it took me a while to figure out what the distinction was between them.

I've also added the currently unused activeStoresAndAtomicsMask() method.
While I'm not keen on adding dead code, the other two methods make reference to the fact this is the method you probably want to use if you are doing something that has side effects.
Given that it's inline, it should have no bloat to the final binary while it is currently unused.

Bug: b/140294254
Change-Id: I0f1517aef86679f277d998f2a3fb301cf657f03a