Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Thanks for discovering the root cause of the slowdown. I don't think it's worth optimizing at this time, though. Papering over it by making the constant data persistent just hides the fact that this should really be precomputed statically, or on the fly.

Currently, `block_size_descriptor` is a whopping 3.3 MB. Times 24 formats, that's 80 MB. Yet this is meant to be implemented in mobile hardware, per texture unit. The first issue is that this structure contains a lot of data meant for the encoder, not the decoder. The bulk of the fat is in `partition_info`, and trimming that off reduces `block_size_descriptor` to 700 KB. Next, observe that `uint8_t partition_of_texel[]` actually only stores 2 bits per element. It's compute by a very complex function. But we really only need two functions, one to compute each bit, which is how this would mostly likely be implemented in hardware. It's clear that ARM's code is only meant as a reference for the theoretical models behind this compression scheme, not as an efficient decoder implementation. Furthermore, the code defines MAX_TEXELS_PER_BLOCK to be 216 (which is 6x6x6) and uses that as the size of many tables, while most formats have much lower texel per block counts.

Taking an even bigger step back: ASTC support is only needed for OpenGL ES 3.2 conformance.  We have yet to complete SwANGLE 3.0 work. So this is a fairly big distraction and could have just been a comment on b/150130101 instead.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1