Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm all for this change, but it does mean that we should take a look at places where DABORT is used to make sure that we are actually able to continue running afterwards. Often times, cases where one aborts means it's not safe to keep executing afterwards, and will likely crash anyway, or worse, not crash and do funny things.

We only use DABORT() in VkSemaphoreExternalFuchsia.hpp, which previously used ABORT() too eargerly: https://swiftshader-review.googlesource.com/c/SwiftShader/+/44018/4/src/Vulkan/VkSemaphoreExternalFuchsia.hpp

I completely agree that when there's a security concern to continue after an unexpected event, and there's no alternative, we should abort. But as described in the commit message and the tracker bug, this can be done using ::abort() instead of this macro which looked too innocuous among its peers.

Also note that security and Vulkan are pretty much antonyms. All memory is under application control, and validation layers are not meant to be used in production. To my knowledge, the spec also never defines a condition for which it requires the process to be terminated. Instead, there are many cases of undefined behavior where it explicitly does not allow program termination.

Anyway, I'm happy to bring back this macro in one form or another when there's a clear need for it.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>