Fix JIT on separate thread
When JIT_IN_SEPARATE_THREAD is defined we generate Reactor routines on
LLVM in a new thread. With Variable::unmaterializedVariables now a
thread-local pointer, we have to avoid touching it from the separate
thread. This is accomplished by moving the createRet() call out of the
lambda which gets called on the separate thread.

Bug: b/153803432
Change-Id: Ic3940e5c52a3009e48fb1e1ff6137db6fd5ca96b