Vulkan/Debug: Don't step for column updates
OpLine is produced for any line / column change.

Column information may be useful for stack frame information (say two function calls on the same line), diagnosing a crash in a complex expression, or some other tooling. However, it's not so useful for IDE debugger stepping, unless you have a location for the span end (which we don't).

Bug: b/155390706
Bug: b/148401179
Change-Id: Ic416c6e3c47044d707bfa36112e1153588669424