Remove typeinfo from GLES exported symbols
This typeinfo was added for UBSan failures reported in crbug.com/737384,
but it didn't address other 'Incorrect-function-pointer-type' failures
for crbug.com/746914. We've suppressed sanitization for these functions
with the NO_SANITIZE_FUNCTION macro, refactored interfaces used across
library boundaries into purely abstract ones for crbug.com/732667, and
use '-mllvm -asan-use-private-alias=1' to address b/128551743.

It's unclear whether this typeinfo still serves ay purpose We shouldn't
have to export typeinfo for objects used only inside the shared library,
to avoid ODR violations.

Remove it entirely for now. If this causes a regression and we still
have to export any typeinfo, we should do so more selectively. For
example to only export typeinfo for objects in the egl namespace:
_ZTS*egl*;
_ZTI*egl*;
Alternatively, or in addition, we can try using
'-mllvm -asan-use-private-alias=1' for more build systems.

Bug: b/155441530
Change-Id: Ia966c40dfe45817f356d11725910afef1bb94d6e