Remove typeinfo from Vulkan exported symbols
Exporting all typeinfo was copied over from the GLES version scripts
which addressed crbug.com/737384. However, this can lead to ODR
violations, and we have several other solutions in place for avoiding
UBSan failures.

Remove it entirely for now. If this causes a regression and we still
have to export some typeinfo, we should do so selectively. For example
to only export typeinfo for objects in the vk namespace:
_ZTS*vk*;
_ZTI*vk*;
Alternatively, or in addition, we can try using
'-mllvm -asan-use-private-alias=1'.

Bug: b/155441530
Change-Id: I07303ac01b05d02b879fff829b6a9945c16fdf82