Fix not checking lower bound of memory range
Through integer overflow, and potentially other means, the dynamic
offset of a Pointer object can become a negative number. The
resulting static + dynamic offset can then be negative, which means
that the combined offset is pointing to some memory before the
beginning of the memory range.

A check was added to make sure the offsets aren't negative. Negative
offsets effectively mean that the pointer is out of bounds.

Bug: b/155862459
Change-Id: I02d4bc2639f3410aeab07fac9f143ee915dc8f4e