Silence warnings on dedicated external memory allocations
The VK_STRUCTURE_TYPE_MEMORY_DEDICATED_ALLOCATE_INFO extension structure
is safe to ignore for Linux and Fuchsia.

We already ignored it for private allocations (see vkAllocateMemory).
Some drivers can do a better job of allocating if they know that a
VkDeviceMemory will contain exactly one object with known properties.
This is the case for NVIDIA, but not SwiftShader.

For external memory, on some platforms (including Android) the shareable
object isn't just a flat blob of memory, it carries format/dimensions
etc. The AHardwareBufferExternalMemory implementation in
VkDeviceMemoryExternalAndroid.hpp obtains the image/buffer from the
VkMemoryDedicatedAllocateInfo structure to passs it to gralloc.

External memory on Linux and Fuchsia are just flat memory allocations,
so on these platforms we can ignore the dedicated allocation info.

Bug: b/156200449
Change-Id: Ia7838bb8a39aac2a1908f314d146ad11f1be79d2