Eliminate the array size from rr::Variable
We only need to know the array size during materialization of Array<>
types, so store it only for that class and make the allocation a
virtual method.

This change removes the REACTOR_MATERIALIZE_LVALUES_ON_DEFINITION
option, since it requires knowing the full size of Variable at
construction, which would require Array<> to still pass that into the
Variable constructor. If we need an debugging feature like this again,
we could also materialize variables on first use instead of at
construction.

Bug: b/155302798
Change-Id: I65b02c1258643dcbd3ae18f3f097cb815efa0830