Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1: Code-Review+2
> 
> (1 comment)
> 
> LGTM, but I'm still very curious whether we can have a more foolproof way of updating the borders. The legacy GLES code had 'lock()' methods which took a type parameter to indicate whether we'd be writing to the images. I think we could do something similar for Vulkan, so we can address b/146157683 fully.

To do that, we'd need some notion of "dirtiness" mechanism in vk::Image, so that we can know when we need the image updated. It might be more optimal to do it that way, since we could potentially write to the image multiple times before updating the borders.

Patch-set: 1
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>